chore(server): fix some major sec issues & finish file structure overhaul
