+++ apps/app/server/trpc/router/analysis/index.ts
@@ -1,0 +1,6 @@
+import { router } from '../../trpc';
+import getUrls from './getUrls';
+
+export const analysisRouter = router({
+  getUrls,
+});
