--- apps/app/server/trpc/router/clip.ts
@@ -1,159 +1,0 @@
-import { z } from 'zod';
-import { ClipSchema } from '@utils/zod/clip';
-
-import { protectedProcedure, router } from '../trpc';
-import { TRPCError } from '@trpc/server';
-import { hasPerms, Perms } from '@server/utils/hasPerms';
-import { serverSanitize } from '@utils/sanitize';
-import * as Sentry from '@sentry/nextjs';
-export const clipRouter = router({
-  get: protectedProcedure
-    .meta({ openapi: { method: 'GET', path: '/clip' } })
-    .input(
-      z
-        .object({
-          clipId: z.string().cuid().optional(),
-        })
-        .transform(input => {
-          return {
-            clipId:
-              input.clipId === undefined
-                ? input.clipId
-                : serverSanitize(input.clipId),
-          };
-        }),
-    )
-    .output(ClipSchema)
-    .query(async ({ input, ctx }) => {
-      if (
-        !hasPerms({
-          userId: ctx.session.user.id,
-          userRole: ctx.session.user.role,
-          requiredPerms: Perms.roleMod,
-          // when this api is used check for owner
-          // itemOwnerId: clip.footage.userId,
-          blacklisted: ctx.session.user.blacklisted,
-        })
-      )
-        throw new TRPCError({
-          code: 'UNAUTHORIZED',
-        });
-
-      try {
-        const clip = await ctx.prisma.clip.findUnique({
-          where: {
-            id: input.clipId,
-          },
-          include: {
-            gameplay: true,
-          },
-        });
-
-        // so the trpc error is thrown
-        if (clip === null) throw new Error('no clip');
-
-        return clip;
-      } catch (error) {
-        Sentry.captureException(error);
-        throw new TRPCError({
-          message: 'No clip with the provided id could be found.',
-          code: 'NOT_FOUND',
-        });
-      }
-    }),
-  delete: protectedProcedure
-    .meta({ openapi: { method: 'DELETE', path: '/clip' } })
-    .input(
-      z
-        .object({
-          clipId: z.string().cuid().optional(),
-        })
-        .transform(input => {
-          return {
-            clipId:
-              input.clipId === undefined
-                ? input.clipId
-                : serverSanitize(input.clipId),
-          };
-        }),
-    )
-    .output(z.object({ message: z.string() }))
-    .mutation(async ({ input, ctx }) => {
-      if (
-        !hasPerms({
-          userId: ctx.session.user.id,
-          userRole: ctx.session.user.role,
-          requiredPerms: Perms.roleMod,
-          // when this api is used check for owner
-          // itemOwnerId: clip.footage.userId,
-          blacklisted: ctx.session.user.blacklisted,
-        })
-      )
-        throw new TRPCError({
-          code: 'UNAUTHORIZED',
-        });
-
-      try {
-        const result = await ctx.prisma.clip.delete({
-          where: {
-            id: input.clipId,
-          },
-        });
-        return {
-          message: `Clip with cuid: ${result.id} was deleted successfully.`,
-        };
-      } catch (error) {
-        Sentry.captureException(error);
-        throw new TRPCError({
-          message: 'No clip with the provided id could be found.',
-          code: 'NOT_FOUND',
-        });
-      }
-    }),
-  create: protectedProcedure
-    .meta({ openapi: { method: 'POST', path: '/clip' } })
-    .input(
-      z
-        .object({
-          gameplayId: z.string().cuid(),
-        })
-        .transform(input => {
-          return {
-            gameplayId: serverSanitize(input.gameplayId),
-          };
-        }),
-    )
-    .output(ClipSchema)
-    .mutation(async ({ input, ctx }) => {
-      if (
-        !hasPerms({
-          userId: ctx.session.user.id,
-          userRole: ctx.session.user.role,
-          requiredPerms: Perms.roleMod,
-          blacklisted: ctx.session.user.blacklisted,
-        })
-      )
-        throw new TRPCError({
-          code: 'UNAUTHORIZED',
-        });
-
-      try {
-        const clip = await ctx.prisma.clip.create({
-          data: {
-            gameplay: {
-              connect: {
-                id: input.gameplayId,
-              },
-            },
-          },
-        });
-        return clip;
-      } catch (error) {
-        Sentry.captureException(error);
-        throw new TRPCError({
-          message: `Unable to create a clip.`,
-          code: 'INTERNAL_SERVER_ERROR',
-        });
-      }
-    }),
-});
