+++ apps/app/server/trpc/router/clip/index.ts
@@ -1,0 +1,10 @@
+import { router } from '../../trpc';
+import create from './create';
+import deleteRoute from './delete';
+import get from './get';
+
+export const clipRouter = router({
+  create,
+  delete: deleteRoute,
+  get,
+});
