+++ apps/app/server/trpc/router/user/index.ts
@@ -1,0 +1,7 @@
+import { router } from '../../trpc';
+import blacklist from './blacklist';
+import deleteRoute from './delete';
+export const userRouter = router({
+  blacklist,
+  delete: deleteRoute,
+});
